requirements.txt
asgiref==3.6.0
asttokens==2.2.1
colorama==0.4.6
Django==4.1.7
django_debug_toolbar==3.8.1
django-jazzmin==2.6.0
django-qrcode==0.3
django-simple-captcha==0.5.17
djangorestframework==3.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
typing_extensions==4.5.0
executing==1.2.0
flake8==6.0.0
icecream==2.1.3
mccabe==0.7.0
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7







